app/requirements.txt
pyautogui
keyboard
openai
python-dotenv